feat(database): split column header component file
