+++ packages/blocks/src/database-block/consts.ts
@@ -1,0 +1,8 @@
+/** column default width */
+export const DEFAULT_COLUMN_WIDTH = 200;
+/** column min width */
+export const DEFAULT_COLUMN_MIN_WIDTH = 100;
+/** column title height */
+export const DEFAULT_COLUMN_TITLE_HEIGHT = 40;
+/** column title height */
+export const DEFAULT_ADD_BUTTON_WIDTH = 40;
